PR-01: extract constants.rs, errors.rs, environ.rs
Behaviour-preserving relocation of self-contained items from main.rs:
- constants.rs: opening DBUS_*/GNOME_FOCUS_*/KDE_*/LOGIND_* block plus
  GNOME extension paths, DCONF_FOCUS_ONLY_KEY, and the GNOME_SHELL_*/
  DBUS_ERROR_* cross-cluster set.
- errors.rs: type DynError.
- environ.rs: Environment, RunOutcome, SessionKind, DesktopFlavor,
  BackendKind, RuntimeTarget, DesktopCapabilities, LifecycleSnapshot,
  impl Environment, plus 8 free functions (session_type_to_session_kind
  through detect_environment).

main.rs: 7983 -> 7743 LOC. Added mod declarations, use blocks, and
#[cfg(test)] re-exports so tests.rs / integration_tests.rs continue
to use super::*.

PR-01-D01 resolved by renaming the module from env to environ
(avoids std::env shadow at 5 call sites in main.rs); plan prose
updated to match.

Verification: cargo build clean; cargo test --bin kanata-switcher
261 passed / 0 failed.
