+++ src/daemon/errors.rs
@@ -1,0 +1,1 @@
+pub(crate) type DynError = Box<dyn std::error::Error + Send + Sync>;
